% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_rd}
\alias{find_rd}
\title{Searches plain Rd in R source files}
\usage{
find_rd()
}
\value{
A tidy data frame that has a \code{\link[=print]{print()}} method that also interacts with the RStudio API.
}
\description{
Looks for Rd-like code in roxygen2 comments, especially useful after running \code{\link[=roxygen2md]{roxygen2md()}}.
This function is designed for interactive use.
}
\examples{
if (interactive()) {
  # Show remaining Rd syntax after conversion.
  find_rd()
}
}
