% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2md-package.R
\docType{package}
\name{roxygen2md-package}
\alias{roxygen2md-package}
\title{roxygen2md: 'Roxygen' to 'Markdown'}
\description{
Converts elements of 'roxygen' documentation to 'markdown'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://roxygen2md.r-lib.org/}
  \item \url{https://github.com/r-lib/roxygen2md}
  \item Report bugs at \url{https://github.com/r-lib/roxygen2md/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com}

Other contributors:
\itemize{
  \item Heather Turner [contributor]
}

}
\keyword{internal}
