% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_text}
\alias{transform_text}
\title{Convert a character vector from Rd to Markdown}
\usage{
transform_text(text)
}
\arguments{
\item{text}{A character vector containing \code{.Rd} style annotations.}
}
\value{
The same vector with \code{.Rd} style annotations converted to Markdown
style annotation.
}
\description{
Convert a character vector from Rd to Markdown
}
\examples{

text <- c(
  "Both \\\\emph{italics} and \\\\bold{bold} text.",
  "And here we have \\\\emph{italics} in some text.",
  "This is \\\\bold{bold} text."
)

transform_text(text)

}
