% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md.R
\name{roxygen2md}
\alias{roxygen2md}
\title{Convert from Rd to Markdown in roxygen2 comments}
\usage{
roxygen2md()
}
\value{
List of changed files, invisibly
}
\description{
Performs various substitutions in all \code{.R} files in a package
to make use of the Markdown functionality in \code{roxygen2}.
See \code{\link[=markdownify]{markdownify()}} for details.
Also attempts to enable Markdown support in \code{roxygen2} by adding a field to
\code{DESCRIPTION}.
Carefully examine the results after running this function!
}
